+++ explorer/src/server/index.ts
@@ -3,11 +3,11 @@ import helmet from 'helmet'
 import http from 'http'
 import mime from 'mime-types'
 import cookieSession from 'cookie-session'
-import adminAuth from './middleware/adminAuth'
-import * as controllers from './controllers'
-import { addRequestLogging, logger } from './logging'
+import adminAuth from '../middleware/adminAuth'
+import * as controllers from '../controllers'
+import { addRequestLogging, logger } from '../logging'
 import { bootstrapRealtime } from './realtime'
-import seed from './seed'
+import seed from '../seed'
 
 export const DEFAULT_PORT = parseInt(process.env.SERVER_PORT, 10) || 8080
 export const COOKIE_EXPIRATION_MS = 86400000 // 1 day in ms
